pyproject.toml
[project]
name = "model-inference-engine"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "httpx>=0.27.2",
    "matplotlib>=3.7.0",
    "pillow>=10.4.0",
    "requests>=2.31.0",
    "sentencepiece>=0.2.0",
    "torch==2.3.1",
    "torchvision==0.18.1",
    "transformers>=4.41.0",
    "ultralytics>=8.3.0",
    "uvicorn>=0.38.0",
]
